fix: separate config from secrets
Changed the implementation to store true config options like log level
in an INI config and left only credential/secret related items in the
.env file.

Added an `IR_` prefix to the .env config options which are specified as
environment variables. This was specifically to address a collision with
USERNAME on Windows reported by a user.